+++ src/components/Courses/Courses.jsx
@@ -1,29 +1,23 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
+
+import { getRole } from '../../store/dataFromStore';
 import { BUTTON_NAMES, ROUTES } from '../../constants';
-
-import { getAuthors, getCourses } from '../../store/dataFromStore';
-import { fetchAllCourses } from '../../store/courses/coureses.actions';
-import { fetchAllAuthors } from '../../store/authors/authors.actions';
-
 import { CourseCard } from './components/CourseCard/CourseCard';
 import { SearchBar } from './components/SearchBar/SearchBar';
 import { Button } from '../../common/Button/Button';
+
+import { useGetCoursesAndAuthors } from '../../common/Hooks/useGetCoursesAndAuthors';
 
 import './courses.scss';
 
 export const Courses = () => {
 	const navigate = useNavigate();
-	const dispatch = useDispatch();
-	const coursesList = useSelector(getCourses);
-	const authorsList = useSelector(getAuthors);
+	const role = useSelector(getRole);
+	const isAdmin = role === 'admin';
 	const [filteredCourses, setFilteredCourses] = useState([]);
-
-	useEffect(() => {
-		dispatch(fetchAllCourses());
-		dispatch(fetchAllAuthors());
-	}, [dispatch]);
+	const { coursesList, authorsList } = useGetCoursesAndAuthors();
 
 	useMemo(() => {
 		setFilteredCourses(coursesList);
@@ -49,7 +43,7 @@ export const Courses = () => {
 		<div className='courses'>
 			<div className='coursesControlsWrapper'>
 				<SearchBar onClick={handleClickSearch} />
-				<Button buttonName={BUTTON_NAMES.addCourse} onClick={handleClickAddCourse} />
+				{isAdmin && <Button buttonName={BUTTON_NAMES.addCourse} onClick={handleClickAddCourse} />}
 			</div>
 
 			{filteredCourses.map(({ id, title, description, creationDate, duration, authors }) => (
